requirements.txt
pdfminer.six
python-docx
pandas
scikit-learn
streamlit
nltk
numpy   

